elevate dimensions to store [skip ci]
